+++ packages/bloc/lib/src/bloc_observer.dart
@@ -1,7 +1,7 @@
 import 'package:bloc/bloc.dart';
 import 'package:meta/meta.dart';
 
-/// An interface for observing the behavior of [Bloc]] instances.
+/// An interface for observing the behavior of [Bloc] instances.
 class BlocObserver {
   /// Called whenever a [Bloc] is instantiated.
   /// In many cases, a cubit may be lazily instantiated and
@@ -9,13 +9,20 @@ class BlocObserver {
   /// instance is created.
   @protected
   @mustCallSuper
-  void onCreate(Bloc bloc) {}
+  void onCreate(BlocBase bloc) {}
 
   /// Called whenever an [event] is `added` to any [bloc] with the given [bloc]
   /// and [event].
   @protected
   @mustCallSuper
-  void onEvent(Bloc bloc, Object? event) {}
+  void onEvent(BlocBase bloc, Object? event) {}
+
+  /// Called whenever a [Change] occurs in any [bloc]
+  /// A [change] occurs when a new state is emitted.
+  /// [onChange] is called before a bloc's state has been updated.
+  @protected
+  @mustCallSuper
+  void onChange(BlocBase bloc, Change change) {}
 
   /// Called whenever a transition occurs in any [bloc] with the given [bloc]
   /// and [transition].
@@ -24,14 +31,14 @@ class BlocObserver {
   /// [onTransition] is called before a [bloc]'s state has been updated.
   @protected
   @mustCallSuper
-  void onTransition(Bloc bloc, Transition transition) {}
+  void onTransition(BlocBase bloc, Transition transition) {}
 
   /// Called whenever an [error] is thrown in any [Bloc] or [Cubit].
   /// The [stackTrace] argument may be [StackTrace.empty] if an error
   /// was received without a stack trace.
   @protected
   @mustCallSuper
-  void onError(Bloc bloc, Object error, StackTrace stackTrace) {}
+  void onError(BlocBase bloc, Object error, StackTrace stackTrace) {}
 
   /// Called whenever a [Bloc] is closed.
   /// [onClose] is called just before the [Bloc] is closed
@@ -39,5 +46,5 @@ class BlocObserver {
   /// emit new states.
   @protected
   @mustCallSuper
-  void onClose(Bloc bloc) {}
+  void onClose(BlocBase bloc) {}
 }