+++ packages/bloc_test/test/cubits/sum_cubit.dart
@@ -6,7 +6,7 @@ import 'counter_cubit.dart';
 
 class SumCubit extends Cubit<int> {
   SumCubit(CounterCubit counterCubit) : super(0) {
-    _countSubscription = counterCubit.listen(
+    _countSubscription = counterCubit.stream.listen(
       (count) => emit(state + count),
     );
   }