+++ packages/replay_bloc/test/replay_bloc_test.dart
@@ -72,7 +72,7 @@ void main() {
       test('does nothing when no state changes have occurred', () async {
         final states = <int>[];
         final bloc = CounterBloc();
-        final subscription = bloc.listen(states.add);
+        final subscription = bloc.stream.listen(states.add);
         bloc.undo();
         await bloc.close();
         await subscription.cancel();
@@ -82,7 +82,7 @@ void main() {
       test('does nothing when limit is 0', () async {
         final states = <int>[];
         final bloc = CounterBloc(limit: 0);
-        final subscription = bloc.listen(states.add);
+        final subscription = bloc.stream.listen(states.add);
         bloc.add(Increment());
         await Future<void>.delayed(Duration.zero);
         bloc.undo();
@@ -94,7 +94,7 @@ void main() {
       test('loses history outside of limit', () async {
         final states = <int>[];
         final bloc = CounterBloc(limit: 1);
-        final subscription = bloc.listen(states.add);
+        final subscription = bloc.stream.listen(states.add);
         bloc..add(Increment())..add(Increment());
         await Future<void>.delayed(Duration.zero);
         bloc..undo()..undo();
@@ -106,7 +106,7 @@ void main() {
       test('reverts to initial state', () async {
         final states = <int>[];
         final bloc = CounterBloc();
-        final subscription = bloc.listen(states.add);
+        final subscription = bloc.stream.listen(states.add);
         bloc.add(Increment());
         await Future<void>.delayed(Duration.zero);
         bloc.undo();
@@ -118,7 +118,7 @@ void main() {
       test('reverts to previous state with multiple state changes ', () async {
         final states = <int>[];
         final bloc = CounterBloc();
-        final subscription = bloc.listen(states.add);
+        final subscription = bloc.stream.listen(states.add);
         bloc..add(Increment())..add(Increment());
         await Future<void>.delayed(Duration.zero);
         bloc.undo();
@@ -155,7 +155,7 @@ void main() {
       test('does nothing when no state changes have occurred', () async {
         final states = <int>[];
         final bloc = CounterBloc();
-        final subscription = bloc.listen(states.add);
+        final subscription = bloc.stream.listen(states.add);
         bloc.redo();
         await bloc.close();
         await subscription.cancel();
@@ -165,7 +165,7 @@ void main() {
       test('does nothing when no undos have occurred', () async {
         final states = <int>[];
         final bloc = CounterBloc();
-        final subscription = bloc.listen(states.add);
+        final subscription = bloc.stream.listen(states.add);
         bloc..add(Increment())..add(Increment());
         await Future<void>.delayed(Duration.zero);
         bloc.redo();
@@ -177,7 +177,7 @@ void main() {
       test('works when one undo has occurred', () async {
         final states = <int>[];
         final bloc = CounterBloc();
-        final subscription = bloc.listen(states.add);
+        final subscription = bloc.stream.listen(states.add);
         bloc..add(Increment())..add(Increment());
         await Future<void>.delayed(Duration.zero);
         bloc
@@ -220,7 +220,7 @@ void main() {
       test('does nothing when undos have been exhausted', () async {
         final states = <int>[];
         final bloc = CounterBloc();
-        final subscription = bloc.listen(states.add);
+        final subscription = bloc.stream.listen(states.add);
         bloc..add(Increment())..add(Increment());
         await Future<void>.delayed(Duration.zero);
         bloc
@@ -237,7 +237,7 @@ void main() {
           'followed by a new state change', () async {
         final states = <int>[];
         final bloc = CounterBloc();
-        final subscription = bloc.listen(states.add);
+        final subscription = bloc.stream.listen(states.add);
         bloc..add(Increment())..add(Increment());
         await Future<void>.delayed(Duration.zero);
         bloc
@@ -317,6 +317,283 @@ void main() {
       test('does nothing when no state changes have occurred', () async {
         final states = <int>[];
         final bloc = CounterBlocMixin();
+        final subscription = bloc.stream.listen(states.add);
+        bloc.undo();
+        await bloc.close();
+        await subscription.cancel();
+        expect(states, isEmpty);
+      });
+
+      test('does nothing when limit is 0', () async {
+        final states = <int>[];
+        final bloc = CounterBlocMixin(limit: 0);
+        final subscription = bloc.stream.listen(states.add);
+        bloc.add(Increment());
+        await Future<void>.delayed(Duration.zero);
+        bloc.undo();
+        await bloc.close();
+        await subscription.cancel();
+        expect(states, const <int>[1]);
+      });
+
+      test('loses history outside of limit', () async {
+        final states = <int>[];
+        final bloc = CounterBlocMixin(limit: 1);
+        final subscription = bloc.stream.listen(states.add);
+        bloc..add(Increment())..add(Increment());
+        await Future<void>.delayed(Duration.zero);
+        bloc..undo()..undo();
+        await bloc.close();
+        await subscription.cancel();
+        expect(states, const <int>[1, 2, 1]);
+      });
+
+      test('reverts to initial state', () async {
+        final states = <int>[];
+        final bloc = CounterBlocMixin();
+        final subscription = bloc.stream.listen(states.add);
+        bloc.add(Increment());
+        await Future<void>.delayed(Duration.zero);
+        bloc.undo();
+        await bloc.close();
+        await subscription.cancel();
+        expect(states, const <int>[1, 0]);
+      });
+
+      test('reverts to previous state with multiple state changes ', () async {
+        final states = <int>[];
+        final bloc = CounterBlocMixin();
+        final subscription = bloc.stream.listen(states.add);
+        bloc..add(Increment())..add(Increment());
+        await Future<void>.delayed(Duration.zero);
+        bloc.undo();
+        await bloc.close();
+        await subscription.cancel();
+        expect(states, const <int>[1, 2, 1]);
+      });
+    });
+
+    group('redo', () {
+      test('does nothing when no state changes have occurred', () async {
+        final states = <int>[];
+        final bloc = CounterBlocMixin();
+        final subscription = bloc.stream.listen(states.add);
+        bloc.redo();
+        await bloc.close();
+        await subscription.cancel();
+        expect(states, isEmpty);
+      });
+
+      test('does nothing when no undos have occurred', () async {
+        final states = <int>[];
+        final bloc = CounterBlocMixin();
+        final subscription = bloc.stream.listen(states.add);
+        bloc..add(Increment())..add(Increment());
+        await Future<void>.delayed(Duration.zero);
+        bloc.redo();
+        await bloc.close();
+        await subscription.cancel();
+        expect(states, const <int>[1, 2]);
+      });
+
+      test('works when one undo has occurred', () async {
+        final states = <int>[];
+        final bloc = CounterBlocMixin();
+        final subscription = bloc.stream.listen(states.add);
+        bloc..add(Increment())..add(Increment());
+        await Future<void>.delayed(Duration.zero);
+        bloc
+          ..undo()
+          ..redo();
+        await bloc.close();
+        await subscription.cancel();
+        expect(states, const <int>[1, 2, 1, 2]);
+      });
+
+      test('does nothing when undos have been exhausted', () async {
+        final states = <int>[];
+        final bloc = CounterBlocMixin();
+        final subscription = bloc.stream.listen(states.add);
+        bloc..add(Increment())..add(Increment());
+        await Future<void>.delayed(Duration.zero);
+        bloc
+          ..undo()
+          ..redo()
+          ..redo();
+        await bloc.close();
+        await subscription.cancel();
+        expect(states, const <int>[1, 2, 1, 2]);
+      });
+
+      test(
+          'does nothing when undos has occurred '
+          'followed by a new state change', () async {
+        final states = <int>[];
+        final bloc = CounterBlocMixin();
+        final subscription = bloc.stream.listen(states.add);
+        bloc..add(Increment())..add(Increment());
+        await Future<void>.delayed(Duration.zero);
+        bloc
+          ..undo()
+          ..add(Decrement());
+        await Future<void>.delayed(Duration.zero);
+        bloc.redo();
+        await bloc.close();
+        await subscription.cancel();
+        expect(states, const <int>[1, 2, 1, 0]);
+      });
+    });
+  });
+
+  group('ReplayBloc (legacy)', () {
+    group('undo', () {
+      test('does nothing when no state changes have occurred', () async {
+        final states = <int>[];
+        final bloc = CounterBloc();
+        // ignore: deprecated_member_use
+        final subscription = bloc.listen(states.add);
+        bloc.undo();
+        await bloc.close();
+        await subscription.cancel();
+        expect(states, isEmpty);
+      });
+
+      test('does nothing when limit is 0', () async {
+        final states = <int>[];
+        final bloc = CounterBloc(limit: 0);
+        // ignore: deprecated_member_use
+        final subscription = bloc.listen(states.add);
+        bloc.add(Increment());
+        await Future<void>.delayed(Duration.zero);
+        bloc.undo();
+        await bloc.close();
+        await subscription.cancel();
+        expect(states, const <int>[1]);
+      });
+
+      test('loses history outside of limit', () async {
+        final states = <int>[];
+        final bloc = CounterBloc(limit: 1);
+        // ignore: deprecated_member_use
+        final subscription = bloc.listen(states.add);
+        bloc..add(Increment())..add(Increment());
+        await Future<void>.delayed(Duration.zero);
+        bloc..undo()..undo();
+        await bloc.close();
+        await subscription.cancel();
+        expect(states, const <int>[1, 2, 1]);
+      });
+
+      test('reverts to initial state', () async {
+        final states = <int>[];
+        final bloc = CounterBloc();
+        // ignore: deprecated_member_use
+        final subscription = bloc.listen(states.add);
+        bloc.add(Increment());
+        await Future<void>.delayed(Duration.zero);
+        bloc.undo();
+        await bloc.close();
+        await subscription.cancel();
+        expect(states, const <int>[1, 0]);
+      });
+
+      test('reverts to previous state with multiple state changes ', () async {
+        final states = <int>[];
+        final bloc = CounterBloc();
+        // ignore: deprecated_member_use
+        final subscription = bloc.listen(states.add);
+        bloc..add(Increment())..add(Increment());
+        await Future<void>.delayed(Duration.zero);
+        bloc.undo();
+        await bloc.close();
+        await subscription.cancel();
+        expect(states, const <int>[1, 2, 1]);
+      });
+    });
+
+    group('redo', () {
+      test('does nothing when no state changes have occurred', () async {
+        final states = <int>[];
+        final bloc = CounterBloc();
+        // ignore: deprecated_member_use
+        final subscription = bloc.listen(states.add);
+        bloc.redo();
+        await bloc.close();
+        await subscription.cancel();
+        expect(states, isEmpty);
+      });
+
+      test('does nothing when no undos have occurred', () async {
+        final states = <int>[];
+        final bloc = CounterBloc();
+        // ignore: deprecated_member_use
+        final subscription = bloc.listen(states.add);
+        bloc..add(Increment())..add(Increment());
+        await Future<void>.delayed(Duration.zero);
+        bloc.redo();
+        await bloc.close();
+        await subscription.cancel();
+        expect(states, const <int>[1, 2]);
+      });
+
+      test('works when one undo has occurred', () async {
+        final states = <int>[];
+        final bloc = CounterBloc();
+        // ignore: deprecated_member_use
+        final subscription = bloc.listen(states.add);
+        bloc..add(Increment())..add(Increment());
+        await Future<void>.delayed(Duration.zero);
+        bloc
+          ..undo()
+          ..redo();
+        await bloc.close();
+        await subscription.cancel();
+        expect(states, const <int>[1, 2, 1, 2]);
+      });
+
+      test('does nothing when undos have been exhausted', () async {
+        final states = <int>[];
+        final bloc = CounterBloc();
+        // ignore: deprecated_member_use
+        final subscription = bloc.listen(states.add);
+        bloc..add(Increment())..add(Increment());
+        await Future<void>.delayed(Duration.zero);
+        bloc
+          ..undo()
+          ..redo()
+          ..redo();
+        await bloc.close();
+        await subscription.cancel();
+        expect(states, const <int>[1, 2, 1, 2]);
+      });
+
+      test(
+          'does nothing when undos has occurred '
+          'followed by a new state change', () async {
+        final states = <int>[];
+        final bloc = CounterBloc();
+        // ignore: deprecated_member_use
+        final subscription = bloc.listen(states.add);
+        bloc..add(Increment())..add(Increment());
+        await Future<void>.delayed(Duration.zero);
+        bloc
+          ..undo()
+          ..add(Decrement());
+        await Future<void>.delayed(Duration.zero);
+        bloc.redo();
+        await bloc.close();
+        await subscription.cancel();
+        expect(states, const <int>[1, 2, 1, 0]);
+      });
+    });
+  });
+
+  group('ReplayBlocMixin (legacy)', () {
+    group('undo', () {
+      test('does nothing when no state changes have occurred', () async {
+        final states = <int>[];
+        final bloc = CounterBlocMixin();
+        // ignore: deprecated_member_use
         final subscription = bloc.listen(states.add);
         bloc.undo();
         await bloc.close();
@@ -327,6 +604,7 @@ void main() {
       test('does nothing when limit is 0', () async {
         final states = <int>[];
         final bloc = CounterBlocMixin(limit: 0);
+        // ignore: deprecated_member_use
         final subscription = bloc.listen(states.add);
         bloc.add(Increment());
         await Future<void>.delayed(Duration.zero);
@@ -339,6 +617,7 @@ void main() {
       test('loses history outside of limit', () async {
         final states = <int>[];
         final bloc = CounterBlocMixin(limit: 1);
+        // ignore: deprecated_member_use
         final subscription = bloc.listen(states.add);
         bloc..add(Increment())..add(Increment());
         await Future<void>.delayed(Duration.zero);
@@ -351,6 +630,7 @@ void main() {
       test('reverts to initial state', () async {
         final states = <int>[];
         final bloc = CounterBlocMixin();
+        // ignore: deprecated_member_use
         final subscription = bloc.listen(states.add);
         bloc.add(Increment());
         await Future<void>.delayed(Duration.zero);
@@ -363,6 +643,7 @@ void main() {
       test('reverts to previous state with multiple state changes ', () async {
         final states = <int>[];
         final bloc = CounterBlocMixin();
+        // ignore: deprecated_member_use
         final subscription = bloc.listen(states.add);
         bloc..add(Increment())..add(Increment());
         await Future<void>.delayed(Duration.zero);
@@ -377,6 +658,7 @@ void main() {
       test('does nothing when no state changes have occurred', () async {
         final states = <int>[];
         final bloc = CounterBlocMixin();
+        // ignore: deprecated_member_use
         final subscription = bloc.listen(states.add);
         bloc.redo();
         await bloc.close();
@@ -387,6 +669,7 @@ void main() {
       test('does nothing when no undos have occurred', () async {
         final states = <int>[];
         final bloc = CounterBlocMixin();
+        // ignore: deprecated_member_use
         final subscription = bloc.listen(states.add);
         bloc..add(Increment())..add(Increment());
         await Future<void>.delayed(Duration.zero);
@@ -399,6 +682,7 @@ void main() {
       test('works when one undo has occurred', () async {
         final states = <int>[];
         final bloc = CounterBlocMixin();
+        // ignore: deprecated_member_use
         final subscription = bloc.listen(states.add);
         bloc..add(Increment())..add(Increment());
         await Future<void>.delayed(Duration.zero);
@@ -413,6 +697,7 @@ void main() {
       test('does nothing when undos have been exhausted', () async {
         final states = <int>[];
         final bloc = CounterBlocMixin();
+        // ignore: deprecated_member_use
         final subscription = bloc.listen(states.add);
         bloc..add(Increment())..add(Increment());
         await Future<void>.delayed(Duration.zero);
@@ -430,6 +715,7 @@ void main() {
           'followed by a new state change', () async {
         final states = <int>[];
         final bloc = CounterBlocMixin();
+        // ignore: deprecated_member_use
         final subscription = bloc.listen(states.add);
         bloc..add(Increment())..add(Increment());
         await Future<void>.delayed(Duration.zero);