+++ packages/bloc/CHANGELOG.md
@@ -1,3 +1,18 @@
+# 7.0.0-nullsafety.4
+
+- **BREAKING**: refactor: `Bloc` and `Cubit` extend `BlocBase`
+  - refactor: `void onError(Bloc bloc, Object error, StackTrace stackTrace)` -> `void onError(BlocBase bloc, Object error, StackTrace stackTrace)`
+  - refactor: `void onEvent(Bloc bloc, Object? event)` -> `void onEvent(BlocBase bloc, Object? event)`
+  - refactor: `void onCreate(Bloc bloc)` -> `void onCreate(BlocBase bloc)`
+  - refactor: `void onClose(Bloc bloc)` -> `void onClose(BlocBase bloc)`
+- **BREAKING**: refactor: `Transition` event is required
+- **BREAKING**: refactor: `Bloc` and `Cubit` do not extend `Stream` and implement `Sink`
+  - refactor: use `bloc.stream` or `cubit.stream` to access `Stream<State>`
+    - `myBloc.map(...)` -> `myBloc.stream.map(...)`
+  - refactor: deprecate `bloc.listen` in favor of `bloc.stream.listen`
+- **BREAKING**: revert: refactor: `Change` and `onChange` removed in favor of `Transition` and `onTransition`
+- fix: `toString` on `Transition` excludes null `event`
+
 # 7.0.0-nullsafety.3
 
 - fix: `transformEvents` multiple subscriptions issue